chore: :art: move to src directory
